requirements.txt
python-nmap
python-whois
rich
